+++ release_notes.md
@@ -31,7 +31,7 @@ First release into griffon artifacts portal.
 
 1.  Deprecated toXXX() used by binding converter; use the new numberTextField() node that will create a JFormattedTextField.
 2.  New maskTextField() node that will create a JFormattedTextField with MaskFormatter.
-3.  generate-all will automatically set startupGroups to the generated MVCGroup name.
+3.  generate-all with --set-startup will automatically set startupGroups to the generated MVCGroup name.
 4.  Minor improvements on default templates.
 5.  remove() will try to merge entity with current persistence context if it is not managed.
 6.  Add a boolean parameter, resume, with default value true to beginTransaction().  Setting resume to false will force start a new transaction (no transaction propagation).